+++ mikeio1d/geometry/protocols.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from typing import Protocol
+from shapely.geometry.base import BaseGeometry
+
+
+class ConvertableToShapely(Protocol):
+    def to_shapely(self) -> BaseGeometry:
+        ...